requirements.txt
slack-bolt>=1.18.0
pyzotero>=1.5.0
PyPDF2>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.0
